+++ extensions/renamer/event.ts
@@ -1,6 +1,9 @@
 import type { EnvironmentAllName } from '../../types/beatmap/shared/environment.ts';
 
-const genericType: { [key: number]: string } = {
+/**
+ * Record of Event Type to Name.
+ */
+export const genericTypeMap: { [key: number]: string } = {
    0: 'Back Lasers',
    1: 'Ring Lights',
    2: 'Left Lasers',
@@ -27,7 +30,10 @@ const genericType: { [key: number]: string } = {
    43: 'Special Event 4',
 };
 
-const environmentType: { [env in EnvironmentAllName]?: { [key: number]: string } } = {
+/**
+ * Record of Environment to Type to Name.
+ */
+export const environmentTypeMap: { [env in EnvironmentAllName]?: { [key: number]: string } } = {
    LinkinParkEnvironment: {
       0: 'Lane Light',
       1: 'Ceiling Lights',
@@ -197,6 +203,9 @@ const environmentType: { [env in EnvironmentAllName]?: { [key: number]: string }
    },
 };
 
+/**
+ * Safely retrieve the name of an event type.
+ */
 export function eventTypeRename(type: number, environment?: EnvironmentAllName): string {
-   return environmentType[environment!]?.[type] || genericType[type] || 'Unknown';
+   return environmentTypeMap[environment!]?.[type] || genericTypeMap[type] || 'Unknown';
 }